requirements.txt
greenlet==2.0.2
kafka-python==2.0.2
mysql-connector-python==8.0.33
mysqlclient==2.1.1
protobuf==3.20.3
PyMySQL==1.0.3
python-dotenv==1.0.0
SQLAlchemy==2.0.16
SQLAlchemy-serializer==1.4.1
typing_extensions==4.6.3
